+++ internal/dsp/variance.go
@@ -44,13 +44,13 @@ func vpx_get4x4sse_cs_c(src_ptr *uint8, src_stride int, ref_ptr *uint8, ref_stri
 	)
 	for r = 0; r < 4; r++ {
 		for c = 0; c < 4; c++ {
-			var diff int = int(*(*uint8)(unsafe.Add(unsafe.Pointer(src_ptr), c)) - *(*uint8)(unsafe.Add(unsafe.Pointer(ref_ptr), c)))
+			var diff int = int(*(*uint8)(unsafe.Add(unsafe.Pointer(src_ptr), c))) - int(*(*uint8)(unsafe.Add(unsafe.Pointer(ref_ptr), c)))
 			distortion += diff * diff
 		}
 		src_ptr = (*uint8)(unsafe.Add(unsafe.Pointer(src_ptr), src_stride))
 		ref_ptr = (*uint8)(unsafe.Add(unsafe.Pointer(ref_ptr), ref_stride))
 	}
-	return uint32(distortion)
+	return uint32(int32(distortion))
 }
 func vpx_get_mb_ss_c(src_ptr *int16) uint32 {
 	var (
@@ -71,9 +71,9 @@ func variance(src_ptr *uint8, src_stride int, ref_ptr *uint8, ref_stride int, w 
 	*sse = 0
 	for i = 0; i < h; i++ {
 		for j = 0; j < w; j++ {
-			var diff int = int(*(*uint8)(unsafe.Add(unsafe.Pointer(src_ptr), j)) - *(*uint8)(unsafe.Add(unsafe.Pointer(ref_ptr), j)))
+			var diff int = int(*(*uint8)(unsafe.Add(unsafe.Pointer(src_ptr), j))) - int(*(*uint8)(unsafe.Add(unsafe.Pointer(ref_ptr), j)))
 			*sum += diff
-			*sse += uint32(diff * diff)
+			*sse += uint32(int32(diff * diff))
 		}
 		src_ptr = (*uint8)(unsafe.Add(unsafe.Pointer(src_ptr), src_stride))
 		ref_ptr = (*uint8)(unsafe.Add(unsafe.Pointer(ref_ptr), ref_stride))
@@ -110,7 +110,7 @@ func var_filter_block2d_bil_second_pass(src_ptr *uint16, ref_ptr *uint8, src_pix
 func vpx_variance64x64_c(src_ptr *uint8, src_stride int, ref_ptr *uint8, ref_stride int, sse *uint32) uint32 {
 	var sum int
 	variance(src_ptr, src_stride, ref_ptr, ref_stride, 64, 64, sse, &sum)
-	return *sse - uint32((int64(sum)*int64(sum))/(64*64))
+	return uint32(int32(int(*sse) - int(uint32(int32((int64(sum)*int64(sum))/(64*64))))))
 }
 func vpx_sub_pixel_variance64x64_c(src_ptr *uint8, src_stride int, x_offset int, y_offset int, ref_ptr *uint8, ref_stride int, sse *uint32) uint32 {
 	var (
@@ -135,7 +135,7 @@ func vpx_sub_pixel_avg_variance64x64_c(src_ptr *uint8, src_stride int, x_offset 
 func vpx_variance64x32_c(src_ptr *uint8, src_stride int, ref_ptr *uint8, ref_stride int, sse *uint32) uint32 {
 	var sum int
 	variance(src_ptr, src_stride, ref_ptr, ref_stride, 64, 32, sse, &sum)
-	return *sse - uint32((int64(sum)*int64(sum))/(64*32))
+	return uint32(int32(int(*sse) - int(uint32(int32((int64(sum)*int64(sum))/(64*32))))))
 }
 func vpx_sub_pixel_variance64x32_c(src_ptr *uint8, src_stride int, x_offset int, y_offset int, ref_ptr *uint8, ref_stride int, sse *uint32) uint32 {
 	var (
@@ -160,7 +160,7 @@ func vpx_sub_pixel_avg_variance64x32_c(src_ptr *uint8, src_stride int, x_offset 
 func vpx_variance32x64_c(src_ptr *uint8, src_stride int, ref_ptr *uint8, ref_stride int, sse *uint32) uint32 {
 	var sum int
 	variance(src_ptr, src_stride, ref_ptr, ref_stride, 32, 64, sse, &sum)
-	return *sse - uint32((int64(sum)*int64(sum))/(32*64))
+	return uint32(int32(int(*sse) - int(uint32(int32((int64(sum)*int64(sum))/(32*64))))))
 }
 func vpx_sub_pixel_variance32x64_c(src_ptr *uint8, src_stride int, x_offset int, y_offset int, ref_ptr *uint8, ref_stride int, sse *uint32) uint32 {
 	var (
@@ -185,7 +185,7 @@ func vpx_sub_pixel_avg_variance32x64_c(src_ptr *uint8, src_stride int, x_offset 
 func vpx_variance32x32_c(src_ptr *uint8, src_stride int, ref_ptr *uint8, ref_stride int, sse *uint32) uint32 {
 	var sum int
 	variance(src_ptr, src_stride, ref_ptr, ref_stride, 32, 32, sse, &sum)
-	return *sse - uint32((int64(sum)*int64(sum))/(32*32))
+	return uint32(int32(int(*sse) - int(uint32(int32((int64(sum)*int64(sum))/(32*32))))))
 }
 func vpx_sub_pixel_variance32x32_c(src_ptr *uint8, src_stride int, x_offset int, y_offset int, ref_ptr *uint8, ref_stride int, sse *uint32) uint32 {
 	var (
@@ -210,7 +210,7 @@ func vpx_sub_pixel_avg_variance32x32_c(src_ptr *uint8, src_stride int, x_offset 
 func vpx_variance32x16_c(src_ptr *uint8, src_stride int, ref_ptr *uint8, ref_stride int, sse *uint32) uint32 {
 	var sum int
 	variance(src_ptr, src_stride, ref_ptr, ref_stride, 32, 16, sse, &sum)
-	return *sse - uint32((int64(sum)*int64(sum))/(32*16))
+	return uint32(int32(int(*sse) - int(uint32(int32((int64(sum)*int64(sum))/(32*16))))))
 }
 func vpx_sub_pixel_variance32x16_c(src_ptr *uint8, src_stride int, x_offset int, y_offset int, ref_ptr *uint8, ref_stride int, sse *uint32) uint32 {
 	var (
@@ -235,7 +235,7 @@ func vpx_sub_pixel_avg_variance32x16_c(src_ptr *uint8, src_stride int, x_offset 
 func vpx_variance16x32_c(src_ptr *uint8, src_stride int, ref_ptr *uint8, ref_stride int, sse *uint32) uint32 {
 	var sum int
 	variance(src_ptr, src_stride, ref_ptr, ref_stride, 16, 32, sse, &sum)
-	return *sse - uint32((int64(sum)*int64(sum))/(16*32))
+	return uint32(int32(int(*sse) - int(uint32(int32((int64(sum)*int64(sum))/(16*32))))))
 }
 func vpx_sub_pixel_variance16x32_c(src_ptr *uint8, src_stride int, x_offset int, y_offset int, ref_ptr *uint8, ref_stride int, sse *uint32) uint32 {
 	var (
@@ -260,7 +260,7 @@ func vpx_sub_pixel_avg_variance16x32_c(src_ptr *uint8, src_stride int, x_offset 
 func VpxVariance16x16C(src_ptr *uint8, src_stride int, ref_ptr *uint8, ref_stride int, sse *uint32) uint32 {
 	var sum int
 	variance(src_ptr, src_stride, ref_ptr, ref_stride, 16, 16, sse, &sum)
-	return *sse - uint32((int64(sum)*int64(sum))/(16*16))
+	return uint32(int32(int(*sse) - int(uint32(int32((int64(sum)*int64(sum))/(16*16))))))
 }
 func vpx_sub_pixel_variance16x16_c(src_ptr *uint8, src_stride int, x_offset int, y_offset int, ref_ptr *uint8, ref_stride int, sse *uint32) uint32 {
 	var (
@@ -285,7 +285,7 @@ func vpx_sub_pixel_avg_variance16x16_c(src_ptr *uint8, src_stride int, x_offset 
 func vpx_variance16x8_c(src_ptr *uint8, src_stride int, ref_ptr *uint8, ref_stride int, sse *uint32) uint32 {
 	var sum int
 	variance(src_ptr, src_stride, ref_ptr, ref_stride, 16, 8, sse, &sum)
-	return *sse - uint32((int64(sum)*int64(sum))/(16*8))
+	return uint32(int32(int(*sse) - int(uint32(int32((int64(sum)*int64(sum))/(16*8))))))
 }
 func vpx_sub_pixel_variance16x8_c(src_ptr *uint8, src_stride int, x_offset int, y_offset int, ref_ptr *uint8, ref_stride int, sse *uint32) uint32 {
 	var (
@@ -310,7 +310,7 @@ func vpx_sub_pixel_avg_variance16x8_c(src_ptr *uint8, src_stride int, x_offset i
 func vpx_variance8x16_c(src_ptr *uint8, src_stride int, ref_ptr *uint8, ref_stride int, sse *uint32) uint32 {
 	var sum int
 	variance(src_ptr, src_stride, ref_ptr, ref_stride, 8, 16, sse, &sum)
-	return *sse - uint32((int64(sum)*int64(sum))/(8*16))
+	return uint32(int32(int(*sse) - int(uint32(int32((int64(sum)*int64(sum))/(8*16))))))
 }
 func vpx_sub_pixel_variance8x16_c(src_ptr *uint8, src_stride int, x_offset int, y_offset int, ref_ptr *uint8, ref_stride int, sse *uint32) uint32 {
 	var (
@@ -335,7 +335,7 @@ func vpx_sub_pixel_avg_variance8x16_c(src_ptr *uint8, src_stride int, x_offset i
 func VpxVariance8x8C(src_ptr *uint8, src_stride int, ref_ptr *uint8, ref_stride int, sse *uint32) uint32 {
 	var sum int
 	variance(src_ptr, src_stride, ref_ptr, ref_stride, 8, 8, sse, &sum)
-	return *sse - uint32((int64(sum)*int64(sum))/(8*8))
+	return uint32(int32(int(*sse) - int(uint32(int32((int64(sum)*int64(sum))/(8*8))))))
 }
 func vpx_sub_pixel_variance8x8_c(src_ptr *uint8, src_stride int, x_offset int, y_offset int, ref_ptr *uint8, ref_stride int, sse *uint32) uint32 {
 	var (
@@ -360,7 +360,7 @@ func vpx_sub_pixel_avg_variance8x8_c(src_ptr *uint8, src_stride int, x_offset in
 func vpx_variance8x4_c(src_ptr *uint8, src_stride int, ref_ptr *uint8, ref_stride int, sse *uint32) uint32 {
 	var sum int
 	variance(src_ptr, src_stride, ref_ptr, ref_stride, 8, 4, sse, &sum)
-	return *sse - uint32((int64(sum)*int64(sum))/(8*4))
+	return uint32(int32(int(*sse) - int(uint32(int32((int64(sum)*int64(sum))/(8*4))))))
 }
 func vpx_sub_pixel_variance8x4_c(src_ptr *uint8, src_stride int, x_offset int, y_offset int, ref_ptr *uint8, ref_stride int, sse *uint32) uint32 {
 	var (
@@ -385,7 +385,7 @@ func vpx_sub_pixel_avg_variance8x4_c(src_ptr *uint8, src_stride int, x_offset in
 func vpx_variance4x8_c(src_ptr *uint8, src_stride int, ref_ptr *uint8, ref_stride int, sse *uint32) uint32 {
 	var sum int
 	variance(src_ptr, src_stride, ref_ptr, ref_stride, 4, 8, sse, &sum)
-	return *sse - uint32((int64(sum)*int64(sum))/(4*8))
+	return uint32(int32(int(*sse) - int(uint32(int32((int64(sum)*int64(sum))/(4*8))))))
 }
 func vpx_sub_pixel_variance4x8_c(src_ptr *uint8, src_stride int, x_offset int, y_offset int, ref_ptr *uint8, ref_stride int, sse *uint32) uint32 {
 	var (
@@ -410,7 +410,7 @@ func vpx_sub_pixel_avg_variance4x8_c(src_ptr *uint8, src_stride int, x_offset in
 func VpxVariance4x4C(src_ptr *uint8, src_stride int, ref_ptr *uint8, ref_stride int, sse *uint32) uint32 {
 	var sum int
 	variance(src_ptr, src_stride, ref_ptr, ref_stride, 4, 4, sse, &sum)
-	return *sse - uint32((int64(sum)*int64(sum))/(4*4))
+	return uint32(int32(int(*sse) - int(uint32(int32((int64(sum)*int64(sum))/(4*4))))))
 }
 func vpx_sub_pixel_variance4x4_c(src_ptr *uint8, src_stride int, x_offset int, y_offset int, ref_ptr *uint8, ref_stride int, sse *uint32) uint32 {
 	var (
@@ -465,7 +465,7 @@ func vpx_comp_avg_pred_c(comp_pred *uint8, pred *uint8, width int, height int, r
 	)
 	for i = 0; i < height; i++ {
 		for j = 0; j < width; j++ {
-			var tmp int = int(*(*uint8)(unsafe.Add(unsafe.Pointer(pred), j)) + *(*uint8)(unsafe.Add(unsafe.Pointer(ref), j)))
+			var tmp int = int(*(*uint8)(unsafe.Add(unsafe.Pointer(pred), j))) + int(*(*uint8)(unsafe.Add(unsafe.Pointer(ref), j)))
 			*(*uint8)(unsafe.Add(unsafe.Pointer(comp_pred), j)) = uint8(int8((tmp + (1 << (1 - 1))) >> 1))
 		}
 		comp_pred = (*uint8)(unsafe.Add(unsafe.Pointer(comp_pred), width))
