+++ internal/util/endian_inl_h.go
@@ -4,11 +4,11 @@ import "math"
 
 const LOCAL_GCC_VERSION = 0
 
-func BSwap16(x uint16_t) uint16_t {
-	return (x >> 8) | (x&math.MaxUint8)<<8
+func BSwap16(x uint16) uint16 {
+	return uint16(int16((int(x) >> 8) | (int(x)&math.MaxUint8)<<8))
 }
-func BSwap32(x uint32_t) uint32_t {
-	return (x >> 24) | ((x >> 8) & 0xFF00) | ((x << 8) & 0xFF0000) | x<<24
+func BSwap32(x uint32) uint32 {
+	return uint32(int32((int(x) >> 24) | ((int(x) >> 8) & 0xFF00) | ((int(x) << 8) & 0xFF0000) | int(x)<<24))
 }
 func BSwap64(x uint64) uint64 {
 	x = ((x & 0xFFFFFFFF00000000) >> 32) | (x&math.MaxUint32)<<32
